Refactor Check.define. A new argument, `checker`, was added to it at some point. It was used instead of `self.check` if it was defined.
That's not the way JSON schema is supposed to work. There should be a single `check` function for a single `Checker` instance.

It was causing a bad interface. That ``checker`` function needed to take a ``definition`` argument that wasn't used. When calling ``define`` with a ``checker`` argument, the ``definition`` argument wasn't used. And those checker functions needed to throw exceptions with spurious messages, when semantically it makes more sense to return a boolean and it's responsibility of `jsonschema` raising those exceptions with good messages.

Now the `definition` argument of `check` can be a function that takes a value and returns whether the (a boolean).

Additionally, a `frozen` method has been added for creating `FrozenChecker`s out of `Checker`s.